copiato, va, ma non vanno i test

+++ src/mp/progetto/DiscountStrategy.java
@@ -1,5 +1,5 @@
 package mp.progetto;
 
 public interface DiscountStrategy {
-	double applyDiscount(double originalPrice);
+	int applyDiscount(int originalPrice);
 }
